+++ src/pass/std_nist.rs
@@ -8,6 +8,7 @@ pub const DEFAULT_PASSWORD_MAX_LEN: usize = 128;
 pub const DEFAULT_ALGORITHM: Algorithm = Algorithm::Pbkdf2;
 pub const DEFAULT_LENGTH_CALCULATION: LengthCalculationMethod = LengthCalculationMethod::Characters;
 pub const DEFAULT_SALT_LEN: usize = 16; // In bytes
+pub const DEFAULT_XHMAC_ALGORITHM: HashFunction = HashFunction::Sha512;
 pub const MIN_SALT_LEN: usize = 4; // In bytes
 pub const PASS_MIN_MIN_LEN: usize = 8;
 pub const PASS_MIN_MAX_LEN: usize = 64;