+++ CHANGELOG.md
@@ -7,8 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased] - ReleaseDate
 
-### Changed
+### Added
+- Optional additional HMAC with an external salt before or after hashing the password.
 
+### Changed
 - The output base for HOTP and TOTP must now be valid UTF-8.
 - `LIBREAUTH_OATH_CODE_INVALID_UTF8` has been renamed `LIBREAUTH_OATH_INVALID_UTF8`.
 - The `output_base_len` parameter in the `libreauth_hotp_cfg` and `libreauth_totp_cfg` structures has been dropped.